requirements.txt
flask
flask_limiter
python-dotenv
markdown
pyyaml
numpy
torch
tabulate